# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trindade Massas Operações API",
    "version": "0.1.0",
    "description": "Generated from the Zod schemas the routes validate with. Never edit the artifact by hand: change the schema and regenerate."
  },
  "servers": [
    {
      "url": "https://trindademasas.duckdns.org",
      "description": "Production"
    }
  ],
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "success"
        ],
        "additionalProperties": false
      },
...
      "TextResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "additionalProperties": false
      },
...
      "ProductsResponse": {
        "type": "object",
        "properties": {
          "assai": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Nhoque Kg",
                "Quadrada",
                "Rolo 500",
                "Rolo kg",
                "Rolo 2kg",
                "Lashana",
                "Disco 200g",
                "Disco 400g",
                "Disgo 500g"
              ]
            }
          },
          "normal": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Nhoque 400g",
                "Nhoque Kg",
                "Quadrada"
              ]
            }
          }
        },
        "required": [
          "assai",
          "normal"
        ],
        "additionalProperties": false
      },
...
      "TurnoResponse": {
        "type": "object",
        "properties": {
          "turno": {
            "type": "string",
            "enum": [
              "tarde",
              "noite"
            ]
          }
        },
        "required": [
          "turno"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/reports/products": {
      "get": {
        "summary": "List the products each product-check element type offers",
        "description": "The offer, not a constraint. A report stores any product name -- the server echoes what it is given -- so this is the list a client puts in front of an operator, served here so a second client does not have to duplicate the constants the SPA keeps.",
        "tags": [
          "reports"
        ],
        "responses": {
          "200": {
            "description": "The two lists, one per product-check element type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/categories": {
      "get": {
        "summary": "List the active report categories with their tasks",
        "tags": [
          "reports"
        ],
        "responses": {
          "200": {
            "description": "Active categories, each carrying its active tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports": {
      "get": {
        "summary": "List reports within a period",
        "tags": [
          "reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "yesterday",
                "7days",
                "30days",
                "custom"
              ],
              "default": "today"
            },
            "required": false,
            "name": "period",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "required": false,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Reports in the period, newest first. This endpoint passes no actor, so its items carry no lifecycle flags.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The query parameters failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a report",
        "tags": [
          "reports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body failed validation, or a report already exists for that shift and date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/history": {
      "get": {
        "summary": "List report history with pagination and filters",
        "tags": [
          "reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "required": false,
            "name": "date",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            },
            "required": false,
            "name": "month",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "required": false,
            "name": "userId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100,
              "default": 30
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "History items, each with its server-computed lifecycle flags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The query parameters failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/turno": {
      "get": {
        "summary": "Report the shift the server detects from its own clock",
        "description": "Detected in São Paulo time: 18:30 onwards is the tarde shift, and 06:00 onwards the noite shift. A client must ask rather than compute this, so a device with a wrong clock still agrees with the server.",
        "tags": [
          "reports"
        ],
        "responses": {
          "200": {
            "description": "The detected shift",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/{id}/deactivate": {
      "patch": {
        "summary": "Deactivate a report",
        "description": "Administrador only.",
        "tags": [
          "reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The report was deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "The identifier is not a number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Administrador only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{id}/export": {
      "get": {
        "summary": "Render the WhatsApp text for one report",
        "description": "Formatted on the server in Brazilian Portuguese. A client must fetch this text rather than reimplement the format.",
        "tags": [
          "reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The ready-to-send text",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextResponse"
                }
              }
            }
          },
          "400": {
            "description": "The report does not carry every temperature reading its catalog requires",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/photos/public/{token}": {
      "get": {
        "summary": "Fetch a photo by its public token, without authentication",
        "description": "Documented as implemented, not as intended: this route requires a token of at least sixteen characters while the server issues twelve-hex-character tokens, so it answers 404 in practice. The short route `/p/{token}` is the one that works and is what `publicUrl` points at.",
        "tags": [
          "reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The photo bytes",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No such photo token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}